Finalized study ...
After a talk with my teacher Mark, he argued that If i finished all my TODO's the study is sufficient for learning outcomes. Given my work on the pareto distribution hypothesis test, the study looks good enough!

The feedback was:
"Might have been useful in retrospect to make it more likely that this is Zipf's than Pareto. Just add that to the abstract. Also add that you are interested in PCA and LLMs, but that this study is about the distribution of document lengths; so it is a substudy.
Other than that, your statistical tests, plots and explanations look great!"